+++ src/fmalloc.cpp
@@ -6,4 +6,5 @@
 #include "fmalloc_ops.inc"
 #include "fmalloc_tensor.inc"
 #include "fmalloc_alp.inc"
+#include "fmalloc_ooc.inc"
 #include "fmalloc_api.inc"